Fixed several small issues: Less confirm alerts and more inline messages. Trashcan is now an interception point Trashcan.Read Some text errors Automatic close of mobile menu if not used in 2 seconds.

+++ src/java/org/infoglue/cms/controllers/kernel/impl/simple/ContentController.java
@@ -3446,7 +3446,7 @@ public class ContentController extends BaseController
 	
     private static void markForDeletionRecursive(Content content, Iterator parentIterator, Database db, boolean skipRelationCheck, boolean skipServiceBindings, boolean forceDelete, InfoGluePrincipal infogluePrincipal, Map<ContentVO, List<ReferenceBean>> contactPersons, boolean notifyResponsibleOnReferenceChange) throws ConstraintException, SystemException, Exception
     {
-        if(!skipRelationCheck)
+        if(!skipRelationCheck && !content.getIsDeleted())
         {
 	        List referenceBeanList = RegistryController.getController().getReferencingObjectsForContent(content.getId(), -1, true, false, db);
 			if(referenceBeanList != null && referenceBeanList.size() > 0 && !forceDelete)
